+++ lib/types/weather/forecast_day.dart
@@ -46,10 +46,10 @@ class ForecastDay extends ValueObject<Map<String, dynamic>> {
       minTemp: Temperature(json['day']['mintemp_c'].toInt(), TemperatureUnit.celsius),
       avgTemp: Temperature(json['day']['avgtemp_c'].toInt(), TemperatureUnit.celsius),
       maxWind: WindSpeed(json['day']['maxwind_kph'].toInt(), WindUnit.kmh),
-      totalPrecipitation: Precipitation(json['day']['totalprecip_mm'].toDouble(), PrecipitationUnit.mm),
-      avgHumidity: Humidity(json['day']['avghumidity'].toDouble(), HumidityUnit.relative),
+      totalPrecipitation: Precipitation(json['day']['totalprecip_mm'].toInt(), PrecipitationUnit.mm),
+      avgHumidity: Humidity(json['day']['avghumidity'].toInt(), HumidityUnit.relative),
       condition: Condition.fromJson(json['day']['condition']),
-      uv: UV(json['day']['uv'].toDouble()),
+      uv: UV(json['day']['uv'].toInt()),
     );
   }
